+++ options.h
@@ -66,7 +66,7 @@ struct options_flopsarray {
 	// upper size limit for the benchmark's array
 	unsigned end;
 	// number of calculations to perform in tests
-	unsigned calculations;
+	long long unsigned calculations;
 	// the test array grows linearly by this amount (determines granularity)
 	unsigned step;
 };
